move command registry from ocis to ocis-pkg
