+++ ocis-pkg/command/storageauthbearer.go
