+++ ocis-pkg/command/storagemetadata.go
