+++ ocis-pkg/command/thumbnails.go
